dmd: fix error with MSVC compilation

+++ src/depfile_parser_dmd.cc
@@ -121,7 +121,7 @@ yy11:
 
   }
 
-  return out - filename;
+  return static_cast<int>(out - filename);
 }
 
 bool DepfileParserDMD::Parse(string* content, string* err) {
